refactor: promote raft/internal/transport to raft/transport
The transport package was unnecessarily buried under internal/, limiting
its accessibility. Promoting it to raft/transport removes the restriction
while keeping it co-located with the raft package it serves.

It still has full access to raft/internal/* (registry, server, settings)
since those packages are internal to the raft/ subtree and raft/transport
lives within that subtree.
